Add and Remove Users from Groups
This allows us to add and remove users from groups. This will
not throw an error if you try to add the same user to the same
group multiple times, neither will it throw an error if you try
to remove a user from a group they are not a part of.