Use Env Vars When Able
This ensures that we are pulling our env values from process.env
when available. If you want to see what env vars are needed, you
can go to src/shared/env.ts and see what keys are are pulling from.
If you want to ensure a specific key is set on startup, remove the
default value from the call and it will error.

This also fixes an issue where we were sending 2xx status codes
when we had an error along with adding a Logger for our
system to use