Add Client Functions and Group Join Requests
More complexity! We are now able to request joining a group
by sending the Client an email, sponsor email, and group ID.

A Sponsor needs to be someone in the group that wants to vouch
for this person. TODO: validate sponsor is in group.

We are able to render dynamic pages