Add Landing Page
We now have a `/` page being rendered instead of a 404. This will
be one of a few public pages and this page hopes to drive people
towards requesting access.

We hope to gain insight into this page as we go but for now we have
no metrics being captured and no way of viewing the people that have
requested access other than via the DB