Sign Up, Login, Logout Flow
This commit sets up the User FLow for Signing Up, Logging in, and
Logging Out. This handles errors and form validation in order to
allow for a smooth experience of onboarding.

This also makes sure that the Dashboard is gated by some cookie
authentication and ensures that we remove the cookie if the user
logs out