Allow New Help Item Creation
This allows the creation of help items inside a specific group.
We need to make it so that all help items are screened by an
Admin before going "live" so that the Admin can ensure top quality
requests only are being made. This will also ensure that _someone_
somewhere knows about this and can start prodding people.