Add PSQL to K8s
We are now able to start and run a Postgres instance in Minikube,
connect to it in the API, and run the Migrations to set up the
Database.

We have also tested rebuilding and killing Pods and they all work
as expected with the patch. Notice that we change the Patch to
only target specific names so that we do not override the Deployment
of PSQL or other backing services