Add Secrets
This uses files set in .env files that are ignored on Git in order
to create the secret values to use. This allows for us to set them
without git seeing them.

This also means that each person wanting to run this will need
to get a set and copy it into each dir