Local Ingress and Authorization
We have the ability to run this locally via Minikube and have
the Ingress point at the correct things. We also have added
Authorization and in light of that Roles. We need to write a script
to create our own users and user roles as site-admin so that we can
ensure that we are able to perform any actions we want.